requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
colorlog>=6.7.0
pypresence>=4.3.0